Avalon

+++ Avalon/Avalon.Text/WriteArg.cs
@@ -18,7 +18,7 @@ public class WriteArg : Any
     public virtual long MaxWidth { get; set; }
     public virtual long Base { get; set; }
     public virtual long Case { get; set; }
-    public virtual uint FillChar { get; set; }
+    public virtual long FillChar { get; set; }
     public virtual long ValueCount { get; set; }
     public virtual long Count { get; set; }
 }
